requirements.txt
python-telegram-bot==20.7
fastapi
uvicorn
gradio
numpy
python-dotenv
groq
google-genai